# hackathon/docs/api/api_spec.yaml
openapi: 3.1.0
info:
  title: Gemma Hackathon API
  version: 0.1.0
paths:
  /v1/protocols:
    post:
      summary: Create Protocol
      description: Create a protocol record and initial document entry.
      operationId: create_protocol_v1_protocols_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtocolCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Protocols
      description: List all protocols with pagination.
      operationId: list_protocols_v1_protocols_get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/protocols/upload:
    post:
      summary: Upload Protocol
      description: 'Upload a PDF protocol file and create a protocol record.


        If auto_extract is True, extraction runs asynchronously in the background

        after the response is returned to avoid request timeouts for large PDFs.'
      operationId: upload_protocol_v1_protocols_upload_post
      parameters:
      - name: auto_extract
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Auto Extract
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_protocol_v1_protocols_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/protocols/{protocol_id}/extract:
    post:
      summary: Extract Criteria
      description: 'Trigger extraction of atomic criteria for a protocol.


        Extraction runs asynchronously in the background after the response is returned

        to avoid request timeouts for long documents.'
      operationId: extract_criteria_v1_protocols__protocol_id__extract_post
      parameters:
      - name: protocol_id
        in: path
        required: true
        schema:
          type: string
          title: Protocol Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/protocols/{protocol_id}:
    get:
      summary: Get Protocol
      description: Get protocol details.
      operationId: get_protocol_v1_protocols__protocol_id__get
      parameters:
      - name: protocol_id
        in: path
        required: true
        schema:
          type: string
          title: Protocol Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/protocols/{protocol_id}/criteria:
    get:
      summary: List Criteria
      description: List criteria generated for a protocol.
      operationId: list_criteria_v1_protocols__protocol_id__criteria_get
      parameters:
      - name: protocol_id
        in: path
        required: true
        schema:
          type: string
          title: Protocol Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriteriaListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/criteria/{criterion_id}:
    patch:
      summary: Update Criterion
      description: Update a single criterion or its metadata.
      operationId: update_criterion_v1_criteria__criterion_id__patch
      parameters:
      - name: criterion_id
        in: path
        required: true
        schema:
          type: string
          title: Criterion Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/CriterionUpdateRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionUpdateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/criteria/{criterion_id}/ground:
    post:
      summary: Ground Criterion
      description: Retrieve SNOMED candidates and field mappings for a criterion.
      operationId: ground_criterion_v1_criteria__criterion_id__ground_post
      parameters:
      - name: criterion_id
        in: path
        required: true
        schema:
          type: string
          title: Criterion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroundingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/hitl/feedback:
    post:
      summary: Hitl Feedback
      description: Record HITL feedback for criteria, SNOMED candidates, and field
        mappings.
      operationId: hitl_feedback_v1_hitl_feedback_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/HitlFeedbackRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Hitl Feedback V1 Hitl Feedback Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/criteria/{criterion_id}/edits:
    get:
      summary: List Criterion Edits
      description: List all HITL edits for a criterion.
      operationId: list_criterion_edits_v1_criteria__criterion_id__edits_get
      parameters:
      - name: criterion_id
        in: path
        required: true
        schema:
          type: string
          title: Criterion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HitlEditsListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_upload_protocol_v1_protocols_upload_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_protocol_v1_protocols_upload_post
    CriteriaListResponse:
      properties:
        protocol_id:
          type: string
          title: Protocol Id
        criteria:
          items:
            $ref: '#/components/schemas/CriterionResponse'
          type: array
          title: Criteria
      type: object
      required:
      - protocol_id
      - criteria
      title: CriteriaListResponse
      description: Response payload for listing criteria.
...
    CriterionUpdateRequest:
      properties:
        text:
          anyOf:
          - type: string
          - type: 'null'
          title: Text
        criterion_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Criterion Type
      type: object
      title: CriterionUpdateRequest
      description: Payload for updating a criterion.
    CriterionUpdateResponse:
      properties:
        criterion_id:
          type: string
          title: Criterion Id
        status:
          type: string
          title: Status
        criterion:
          $ref: '#/components/schemas/CriterionResponse'
      type: object
      required:
      - criterion_id
      - status
      - criterion
      title: CriterionUpdateResponse
      description: Response payload after updating a criterion.
    ExtractionResponse:
      properties:
        protocol_id:
          type: string
          title: Protocol Id
        status:
          type: string
          title: Status
        criteria_count:
          type: integer
          title: Criteria Count
      type: object
      required:
      - protocol_id
      - status
      - criteria_count
      title: ExtractionResponse
      description: Response payload for extraction trigger.
...
    GroundingCandidateResponse:
      properties:
        code:
          type: string
          title: Code
        display:
          type: string
          title: Display
        ontology:
          type: string
          title: Ontology
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - code
      - display
      - ontology
      - confidence
      title: GroundingCandidateResponse
      description: Response payload for grounding candidates.
    GroundingResponse:
      properties:
        criterion_id:
          type: string
          title: Criterion Id
        candidates:
          items:
            $ref: '#/components/schemas/GroundingCandidateResponse'
          type: array
          title: Candidates
        field_mapping:
          anyOf:
          - $ref: '#/components/schemas/FieldMappingResponse'
          - type: 'null'
      type: object
      required:
      - criterion_id
      - candidates
      - field_mapping
      title: GroundingResponse
      description: Response payload for grounding a criterion.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HitlEditsListResponse:
      properties:
        criterion_id:
          type: string
          title: Criterion Id
        edits:
          items:
            $ref: '#/components/schemas/HitlEditResponse'
          type: array
          title: Edits
      type: object
      required:
      - criterion_id
      - edits
      title: HitlEditsListResponse
      description: Response for listing HITL edits.
    HitlFeedbackRequest:
      properties:
        criterion_id:
          type: string
          title: Criterion Id
        action:
          $ref: '#/components/schemas/HitlAction'
        snomed_code_added:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Code Added
        snomed_code_removed:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Code Removed
        field_mapping_added:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Mapping Added
        field_mapping_removed:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Mapping Removed
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - criterion_id
      - action
      title: HitlFeedbackRequest
      description: Payload for HITL feedback actions.
    ProtocolCreateRequest:
      properties:
        title:
          type: string
          title: Title
        document_text:
          type: string
          title: Document Text
        nct_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Nct Id
        condition:
          anyOf:
          - type: string
          - type: 'null'
          title: Condition
        phase:
          anyOf:
          - type: string
          - type: 'null'
          title: Phase
      type: object
      required:
      - title
      - document_text
      title: ProtocolCreateRequest
      description: Request payload for creating a protocol entry.
    ProtocolDetailResponse:
      properties:
        protocol_id:
          type: string
          title: Protocol Id
        title:
          type: string
          title: Title
        document_text:
          type: string
          title: Document Text
        nct_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Nct Id
        condition:
          anyOf:
          - type: string
          - type: 'null'
          title: Condition
        phase:
          anyOf:
          - type: string
          - type: 'null'
          title: Phase
        criteria_count:
          type: integer
          title: Criteria Count
      type: object
      required:
      - protocol_id
      - title
      - document_text
      - criteria_count
      title: ProtocolDetailResponse
      description: Response for protocol detail view.
...
    ProtocolListResponse:
      properties:
        protocols:
          items:
            $ref: '#/components/schemas/ProtocolListItem'
          type: array
          title: Protocols
        total:
          type: integer
          title: Total
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - protocols
      - total
      - skip
      - limit
      title: ProtocolListResponse
      description: Response for listing protocols.
    ProtocolResponse:
      properties:
        protocol_id:
          type: string
          title: Protocol Id
        title:
          type: string
          title: Title
      type: object
      required:
      - protocol_id
      - title
      title: ProtocolResponse
      description: Response payload for a created protocol.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError